refactor: move trust broker error payload to dedicated type
